Change VLP(addr) to LP(addr) in DOT spec
Hi @jedisct1 !

Am I right that this is a type and you actually wanted to use `LP(addr)` instead of `VLP(addr)` in the spec?

There is no implementation yet so I have to assume. It's just we're going to use stamps in AG for all types of DNS servers, and this seemed strange.