Add DNS-over-QUIC stamp spec
I guess I proposed it in the [wrong place](https://github.com/DNSCrypt/dnscrypt-proxy/issues/1459) at first.

I suppose it makes sense to have DoQ in the DNS stamps spec since there's a high chance that it will become a standard.

Regarding your comment:

> I'm having a hard time seeing a reason to ever use DoQ, especially since DoH will eventually be DNS-over-HTTP/3.

DoQ looks more "pure" to me. QUIC is just a transport that is used for transferring DNS messages and it won't be used for other purposes. Unlike HTTP, that can be misused in many ways.